Scripts/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
gitdb==4.0.9
GitPython==3.1.41
idna==3.4
requests==2.28.1
smmap==5.0.0
urllib3==1.26.18
############################
#requirements for confluence
atlassian-python-api==3.41.2
Jinja2==3.1.2
pandas
pytz==2023.3
lxml==4.9.3
############################